requirements.txt
flask>=3.0
flask-cors>=4.0
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.1.0
chromadb>=0.4.0
dashscope>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.10
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0
prometheus-client>=0.19.0
tqdm>=4.65.0
